backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0